+++ eslint.config.mjs
@@ -3,7 +3,7 @@ import tseslint from "typescript-eslint";
 
 export default tseslint.config(
   {
-    ignores: [".build/", "coverage/", "node_modules/"],
+    ignores: [".build/", "**/dist/", "coverage/", "node_modules/"],
   },
   eslint.configs.recommended,
   ...tseslint.configs.recommended,